# Backend/src/api/openapi.yaml
openapi: 3.0.0
info:
  title: em-Eukal-App
  version: '1.0'
  description: 'Diese API dient der Em-Eukal Produktbewertungsapp und bietet Endpunkte zum Erstellen, Löschen und Bearbeiten von Produkteinträgen und dazugehöriger Bewertungen. '
  contact:
    name: Pfängerle und Meyrle AG
    email: support@PuMAG.com
    url: PuMAG.com
servers:
  - url: 'http://localhost:3000'
    description: Julia Pfanger
paths:
  /products:
    get:
      summary: Your GET endpoint
      tags: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
      operationId: get-products
      description: Liefert alle Produkteinträge
    post:
      summary: ''
      operationId: post-products
      responses:
        "201":
          description: Created
          headers:
            Location:
              schema:
                type: string
      description: Erstellt neuen Produkteintrag
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Product'
        description: ''
  '/products/{id}':
    parameters:
      - schema:
          type: string
        name: id
        in: path
        required: true
    get:
      summary: Your GET endpoint
      tags: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          description: Not Found
      operationId: get-products-id
      description: Liefert Produkteintrag mit dieser ID
    put:
      summary: ''
      operationId: put-products-id
      responses:
        '200':
          description: OK
        '404':
          description: Not Found
      description: Aktualisiert einen Produkteintrag
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Product'
    patch:
      summary: ''
      operationId: patch-products-id
      responses:
        '200':
          description: OK
        '404':
          description: Not Found
      description: Aktualisiert einzelne Werte des Produkteintrages
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Product'
        description: ''
    delete:
      summary: ''
      operationId: delete-products-id
      responses:
        '200':
          description: OK
        '404':
          description: Not Found
      description: Löscht Produkteintrag
  /reviews:
    get:
      summary: Your GET endpoint
      tags: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Review'
      operationId: get-reviews
      description: Liefert Produktbewertungen
    post:
      summary: ''
      operationId: post-reviews
      responses:
        '201':
          description: Created
          headers:
            Location:
              schema:
                type: string
      description: Erstellt neue Produktbewertung
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Review'
        description: ''
  '/reviews/{id}':
    parameters:
      - schema:
          type: string
        name: id
        in: path
        required: true
    get:
      summary: Your GET endpoint
      tags: [ ]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Review'
        '404':
          description: Not Found
      operationId: get-reviews-id
      description: Liefert einzelne Produktbewertung
    put:
      summary: ''
      operationId: put-reviews-id
      responses:
        '200':
          description: OK
        '404':
          description: Not Found
      description: Aktualisiert eine Produktbewertung
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Review'
    patch:
      summary: ''
      operationId: patch-reviews-id
      responses:
        '200':
          description: OK
        '404':
          description: Not Found
      description: Aktualisiert einzelne Werte einer Produktbewertung
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Review'
    delete:
      summary: ''
      operationId: delete-reviews-id
      responses:
        '200':
          description: OK
        '404':
          description: Not Found
      description: Löscht einzelne Produktbewertung
  '/productReviews/{id}':
    parameters:
      - schema:
          type: string
        name: id
        in: path
        required: true
    get:
      summary: Your GET endpoint
      tags: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  '':
                    type: array
                    items:
                      $ref: '#/components/schemas/Review'
      operationId: get-productReviews-id
      description: Liefert Produktbewertungen zum Produkt der angegebenen ID
components:
  schemas:
    Product:
      title: Product
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        picture_url:
          type: string
        like:
          type: integer
        dislike:
          type: integer
        description:
          type: string
    Review:
      title: Review
      type: object
      properties:
        id:
          type: integer
        product_id:
          type: integer
        name:
          type: string
        comment:
          type: string
        like:
          type: boolean
  securitySchemes: {}